feat: Add support for above 2 GB results.

+++ source/Databricks/source/SqlStatementExecution/Statement/DatabricksStatementResponse.cs
@@ -56,13 +56,13 @@ internal class Manifest
 
     public Schema schema { get; set; }
 
-    public int total_chunk_count { get; set; }
+    public long total_chunk_count { get; set; }
 
     public Chunks[] chunks { get; set; }
 
-    public int total_row_count { get; set; }
+    public long total_row_count { get; set; }
 
-    public int total_byte_count { get; set; }
+    public long total_byte_count { get; set; }
 
     public bool truncated { get; set; }
 }
@@ -91,13 +91,13 @@ internal class Columns
 
 internal class Chunks
 {
-    public int chunk_index { get; set; }
+    public long chunk_index { get; set; }
 
-    public int row_offset { get; set; }
+    public long row_offset { get; set; }
 
-    public int row_count { get; set; }
+    public long row_count { get; set; }
 
-    public int byte_count { get; set; }
+    public long byte_count { get; set; }
 }
 
 internal class Result
@@ -107,13 +107,13 @@ internal class Result
 
 internal class External_links
 {
-    public int chunk_index { get; set; }
+    public long chunk_index { get; set; }
 
-    public int row_offset { get; set; }
+    public long row_offset { get; set; }
 
-    public int row_count { get; set; }
+    public long row_count { get; set; }
 
-    public int byte_count { get; set; }
+    public long byte_count { get; set; }
 
     public string external_link { get; set; }
